refactor: use ES module syntax and bundle by webpack
