+++ index.js
@@ -1,5 +1,9 @@
-module.exports = Object.assign({},
-  require("./lib/webidl2.js"),
-  require("./lib/writer.js"),
-  require("./lib/validator.js")
-);
+import parse from "./lib/webidl2.js";
+import write from "./lib/writer.js";
+import validate from "./lib/validator.js";
+
+export default {
+  parse,
+  write,
+  validate
+};
