+++ lib/writer.js
@@ -1,321 +1,307 @@
 "use strict";
 
-(() => {
-  function noop(arg) {
-    return arg;
+function noop(arg) {
+  return arg;
+}
+
+const templates = {
+  wrap: items => items.join(""),
+  trivia: noop,
+  name: noop,
+  reference: noop,
+  type: noop,
+  inheritance: noop,
+  definition: noop,
+  extendedAttribute: noop,
+  extendedAttributeReference: noop
+};
+
+export default function write(ast, { templates: ts = templates } = {}) {
+  ts = Object.assign({}, templates, ts);
+
+  /**
+   * @param {string[]} strings
+   * @param  {...any} args
+   */
+  function wrap(strings, ...args) {
+    return ts.wrap([].concat(...strings.map((s, i) => [s, args[i] || ""])));
   }
 
-  const templates = {
-    wrap: items => items.join(""),
-    trivia: noop,
-    name: noop,
-    reference: noop,
-    type: noop,
-    inheritance: noop,
-    definition: noop,
-    extendedAttribute: noop,
-    extendedAttributeReference: noop
+  function reference(raw, unescaped) {
+    return ts.reference(raw, unescaped || raw);
+  }
+
+  function extract_trivia(object) {
+    const batch = {};
+    for (const key in object.trivia) {
+      batch[key] = ts.trivia(object.trivia[key]);
+    }
+    return batch;
+  }
+
+  function token(t, value) {
+    return t ? wrap`${ts.trivia(t.trivia)}${value || t.value}` : "";
+  }
+
+  function reference_token(t, unescaped) {
+    return t ? wrap`${ts.trivia(t.trivia)}${reference(t.value, unescaped)}` : "";
+  }
+
+  function name_token(t, arg) {
+    return t ? wrap`${ts.trivia(t.trivia)}${ts.name(t.value, arg)}` : "";
+  }
+
+  function type_body(it) {
+    const trivia = extract_trivia(it);
+    if (it.union) {
+      const subtypes = ts.wrap(it.idlType.map(type));
+      return wrap`${trivia.open}(${subtypes}${trivia.close})`;
+    } else if (it.generic) {
+      const genericName = wrap`${trivia.base}${reference(it.generic.value)}`;
+      const subtypes = ts.wrap(it.idlType.map(type));
+      const gTrivia = extract_trivia(it.generic);
+      return wrap`${genericName}${gTrivia.open}<${subtypes}${gTrivia.close}>`;
+    }
+    if (!it.prefix && !it.postfix) {
+      return wrap`${trivia.base}${reference(it.escapedBaseName, it.baseName)}`;
+    }
+    const precedingTrivia = it.prefix ? ts.trivia(it.prefix.trivia) : trivia.base;
+    const prefix = it.prefix ? wrap`${it.prefix.value}${trivia.base}` : "";
+    const ref = reference(wrap`${prefix}${it.baseName}${token(it.postfix)}`, it.idlType);
+    return wrap`${precedingTrivia}${ref}`;
+  }
+  function type(it) {
+    const ext = extended_attributes(it.extAttrs);
+    const body = type_body(it);
+    const nullable = token(it.nullable, "?");
+    const separator = token(it.separator);
+    return wrap`${ext}${body}${nullable}${separator}`;
+  }
+  function default_(def) {
+    if (!def) {
+      return "";
+    }
+    return ts.wrap([
+      token(def.tokens.assign),
+      ...def.expression.map(t => token(t))
+    ]);
+  }
+  function argument(arg) {
+    return ts.wrap([
+      extended_attributes(arg.extAttrs),
+      token(arg.tokens.optional),
+      ts.type(type(arg.idlType)),
+      token(arg.tokens.variadic),
+      name_token(arg.tokens.name, { data: arg }),
+      default_(arg.default),
+      token(arg.tokens.separator)
+    ]);
+  }
+  function identifier(id) {
+    return wrap`${ts.trivia(id.trivia)}${reference(id.value)}${token(id.separator)}`;
+  }
+  function make_ext_at(it) {
+    const { rhsType } = it.params;
+    return ts.wrap([
+      ts.trivia(it.tokens.name.trivia),
+      ts.extendedAttribute(ts.wrap([
+        ts.extendedAttributeReference(it.name),
+        token(it.params.tokens.assign),
+        reference_token(it.params.tokens.secondaryName),
+        token(it.params.tokens.open),
+        ...!it.params.list ? [] :
+          it.params.list.map(
+            rhsType === "identifier-list" ? identifier : argument
+          ),
+        token(it.params.tokens.close)
+      ])),
+      token(it.tokens.separator)
+    ]);
+  }
+  function extended_attributes(eats) {
+    if (!eats) return "";
+    return ts.wrap([
+      token(eats.tokens.open),
+      ...eats.items.map(make_ext_at),
+      token(eats.tokens.close)
+    ]);
+  }
+
+  function operation(it, parent) {
+    const body = it.body ? [
+      ts.type(type(it.body.idlType)),
+      name_token(it.body.tokens.name, { data: it, parent }),
+      token(it.body.tokens.open),
+      ts.wrap(it.body.arguments.map(argument)),
+      token(it.body.tokens.close),
+    ] : [];
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.special),
+      ...body,
+      token(it.tokens.termination)
+    ]), { data: it, parent });
+  }
+
+  function attribute(it, parent) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.special),
+      token(it.tokens.readonly),
+      token(it.tokens.base),
+      ts.type(type(it.idlType)),
+      name_token(it.tokens.name, { data: it, parent }),
+      token(it.tokens.termination)
+    ]), { data: it, parent });
+  }
+
+  function inheritance(inh) {
+    if (!inh) {
+      return "";
+    }
+    return ts.wrap([
+      token(inh.tokens.colon),
+      ts.trivia(inh.tokens.name.trivia),
+      ts.inheritance(reference(inh.escapedName, inh.name))
+    ]);
+  }
+
+  function container(it) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.partial),
+      token(it.tokens.base),
+      token(it.tokens.mixin),
+      name_token(it.tokens.name, { data: it }),
+      inheritance(it.inheritance),
+      token(it.tokens.open),
+      iterate(it.members, it),
+      token(it.tokens.close),
+      token(it.tokens.termination)
+    ]), { data: it });
+  }
+
+  function field(it, parent) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.required),
+      ts.type(type(it.idlType)),
+      name_token(it.tokens.name, { data: it, parent }),
+      default_(it.default),
+      token(it.tokens.termination)
+    ]), { data: it, parent });
+  }
+  function const_(it, parent) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.base),
+      ts.type(type(it.idlType)),
+      name_token(it.tokens.name, { data: it, parent }),
+      token(it.tokens.assign),
+      token(it.tokens.value),
+      token(it.tokens.termination)
+    ]), { data: it, parent });
+  }
+  function typedef(it) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.base),
+      ts.type(type(it.idlType)),
+      name_token(it.tokens.name, { data: it }),
+      token(it.tokens.termination)
+    ]), { data: it });
+  }
+  function includes(it) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      reference_token(it.tokens.target, it.target),
+      token(it.tokens.includes),
+      reference_token(it.tokens.mixin, it.includes),
+      token(it.tokens.termination)
+    ]), { data: it });
+  }
+  function callback(it) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.base),
+      name_token(it.tokens.name, { data: it }),
+      token(it.tokens.assign),
+      ts.type(type(it.idlType)),
+      token(it.tokens.open),
+      ...it.arguments.map(argument),
+      token(it.tokens.close),
+      token(it.tokens.termination),
+    ]), { data: it });
+  }
+  function enum_(it) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.base),
+      name_token(it.tokens.name, { data: it }),
+      token(it.tokens.open),
+      iterate(it.values, it),
+      token(it.tokens.close),
+      token(it.tokens.termination)
+    ]), { data: it });
+  }
+  function enum_value(v, parent) {
+    return wrap`${ts.trivia(v.trivia)}${
+      ts.definition(wrap`"${ts.name(v.value, { data: v, parent })}"`, { data: v, parent })
+    }${token(v.separator)}`;
+  }
+  function iterable_like(it, parent) {
+    return ts.definition(ts.wrap([
+      extended_attributes(it.extAttrs),
+      token(it.tokens.readonly),
+      token(it.tokens.base),
+      token(it.tokens.open),
+      ts.wrap(it.idlType.map(type)),
+      token(it.tokens.close),
+      token(it.tokens.termination)
+    ]), { data: it, parent });
+  }
+  function callbackInterface(it) {
+    return ts.definition(
+      wrap`${ts.trivia(it.trivia.callback)}callback${container(it)}`,
+      { data: it }
+    );
+  }
+  function eof(it) {
+    return ts.trivia(it.trivia);
+  }
+
+  const table = {
+    interface: container,
+    "interface mixin": container,
+    namespace: container,
+    operation,
+    attribute,
+    dictionary: container,
+    field,
+    const: const_,
+    typedef,
+    includes,
+    callback,
+    enum: enum_,
+    "enum-value": enum_value,
+    iterable: iterable_like,
+    legacyiterable: iterable_like,
+    maplike: iterable_like,
+    setlike: iterable_like,
+    "callback interface": callbackInterface,
+    eof
   };
-
-  function write(ast, { templates: ts = templates } = {}) {
-    ts = Object.assign({}, templates, ts);
-
-    /**
-     * @param {string[]} strings
-     * @param  {...any} args
-     */
-    function wrap(strings, ...args) {
-      return ts.wrap([].concat(...strings.map((s, i) => [s, args[i] || ""])));
+  function dispatch(it, parent) {
+    const dispatcher = table[it.type];
+    if (!dispatcher) {
+      throw new Error(`Type "${it.type}" is unsupported`);
     }
-
-    function reference(raw, unescaped) {
-      return ts.reference(raw, unescaped || raw);
-    }
-
-    function extract_trivia(object) {
-      const batch = {};
-      for (const key in object.trivia) {
-        batch[key] = ts.trivia(object.trivia[key]);
-      }
-      return batch;
-    }
-
-    function token(t, value) {
-      return t ? wrap`${ts.trivia(t.trivia)}${value || t.value}` : "";
-    }
-
-    function reference_token(t, unescaped) {
-      return t ? wrap`${ts.trivia(t.trivia)}${reference(t.value, unescaped)}` : "";
-    }
-
-    function name_token(t, arg) {
-      return t ? wrap`${ts.trivia(t.trivia)}${ts.name(t.value, arg)}` : "";
-    }
-
-    function type_body(it) {
-      const trivia = extract_trivia(it);
-      if (it.union) {
-        const subtypes = ts.wrap(it.idlType.map(type));
-        return wrap`${trivia.open}(${subtypes}${trivia.close})`;
-      } else if (it.generic) {
-        const genericName = wrap`${trivia.base}${reference(it.generic.value)}`;
-        const subtypes = ts.wrap(it.idlType.map(type));
-        const gTrivia = extract_trivia(it.generic);
-        return wrap`${genericName}${gTrivia.open}<${subtypes}${gTrivia.close}>`;
-      }
-      if (!it.prefix && !it.postfix) {
-        return wrap`${trivia.base}${reference(it.escapedBaseName, it.baseName)}`;
-      }
-      const precedingTrivia = it.prefix ? ts.trivia(it.prefix.trivia) : trivia.base;
-      const prefix = it.prefix ? wrap`${it.prefix.value}${trivia.base}` : "";
-      const ref = reference(wrap`${prefix}${it.baseName}${token(it.postfix)}`, it.idlType);
-      return wrap`${precedingTrivia}${ref}`;
-    }
-    function type(it) {
-      const ext = extended_attributes(it.extAttrs);
-      const body = type_body(it);
-      const nullable = token(it.nullable, "?");
-      const separator = token(it.separator);
-      return wrap`${ext}${body}${nullable}${separator}`;
-    }
-    function default_(def) {
-      if (!def) {
-        return "";
-      }
-      return ts.wrap([
-        token(def.tokens.assign),
-        ...def.expression.map(t => token(t))
-      ]);
-    }
-    function argument(arg) {
-      return ts.wrap([
-        extended_attributes(arg.extAttrs),
-        token(arg.tokens.optional),
-        ts.type(type(arg.idlType)),
-        token(arg.tokens.variadic),
-        name_token(arg.tokens.name, { data: arg }),
-        default_(arg.default),
-        token(arg.tokens.separator)
-      ]);
-    }
-    function identifier(id) {
-      return wrap`${ts.trivia(id.trivia)}${reference(id.value)}${token(id.separator)}`;
-    }
-    function make_ext_at(it) {
-      const { rhsType } = it.params;
-      return ts.wrap([
-        ts.trivia(it.tokens.name.trivia),
-        ts.extendedAttribute(ts.wrap([
-          ts.extendedAttributeReference(it.name),
-          token(it.params.tokens.assign),
-          reference_token(it.params.tokens.secondaryName),
-          token(it.params.tokens.open),
-          ...!it.params.list ? [] :
-            it.params.list.map(
-              rhsType === "identifier-list" ? identifier : argument
-            ),
-          token(it.params.tokens.close)
-        ])),
-        token(it.tokens.separator)
-      ]);
-    }
-    function extended_attributes(eats) {
-      if (!eats) return "";
-      return ts.wrap([
-        token(eats.tokens.open),
-        ...eats.items.map(make_ext_at),
-        token(eats.tokens.close)
-      ]);
-    }
-
-    function operation(it, parent) {
-      const body = it.body ? [
-        ts.type(type(it.body.idlType)),
-        name_token(it.body.tokens.name, { data: it, parent }),
-        token(it.body.tokens.open),
-        ts.wrap(it.body.arguments.map(argument)),
-        token(it.body.tokens.close),
-      ] : [];
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.special),
-        ...body,
-        token(it.tokens.termination)
-      ]), { data: it, parent });
-    }
-
-    function attribute(it, parent) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.special),
-        token(it.tokens.readonly),
-        token(it.tokens.base),
-        ts.type(type(it.idlType)),
-        name_token(it.tokens.name, { data: it, parent }),
-        token(it.tokens.termination)
-      ]), { data: it, parent });
-    }
-
-    function inheritance(inh) {
-      if (!inh) {
-        return "";
-      }
-      return ts.wrap([
-        token(inh.tokens.colon),
-        ts.trivia(inh.tokens.name.trivia),
-        ts.inheritance(reference(inh.escapedName, inh.name))
-      ]);
-    }
-
-    function container(it) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.partial),
-        token(it.tokens.base),
-        token(it.tokens.mixin),
-        name_token(it.tokens.name, { data: it }),
-        inheritance(it.inheritance),
-        token(it.tokens.open),
-        iterate(it.members, it),
-        token(it.tokens.close),
-        token(it.tokens.termination)
-      ]), { data: it });
-    }
-
-    function field(it, parent) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.required),
-        ts.type(type(it.idlType)),
-        name_token(it.tokens.name, { data: it, parent }),
-        default_(it.default),
-        token(it.tokens.termination)
-      ]), { data: it, parent });
-    }
-    function const_(it, parent) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.base),
-        ts.type(type(it.idlType)),
-        name_token(it.tokens.name, { data: it, parent }),
-        token(it.tokens.assign),
-        token(it.tokens.value),
-        token(it.tokens.termination)
-      ]), { data: it, parent });
-    }
-    function typedef(it) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.base),
-        ts.type(type(it.idlType)),
-        name_token(it.tokens.name, { data: it }),
-        token(it.tokens.termination)
-      ]), { data: it });
-    }
-    function includes(it) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        reference_token(it.tokens.target, it.target),
-        token(it.tokens.includes),
-        reference_token(it.tokens.mixin, it.includes),
-        token(it.tokens.termination)
-      ]), { data: it });
-    }
-    function callback(it) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.base),
-        name_token(it.tokens.name, { data: it }),
-        token(it.tokens.assign),
-        ts.type(type(it.idlType)),
-        token(it.tokens.open),
-        ...it.arguments.map(argument),
-        token(it.tokens.close),
-        token(it.tokens.termination),
-      ]), { data: it });
-    }
-    function enum_(it) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.base),
-        name_token(it.tokens.name, { data: it }),
-        token(it.tokens.open),
-        iterate(it.values, it),
-        token(it.tokens.close),
-        token(it.tokens.termination)
-      ]), { data: it });
-    }
-    function enum_value(v, parent) {
-      return wrap`${ts.trivia(v.trivia)}${
-        ts.definition(wrap`"${ts.name(v.value, { data: v, parent })}"`, { data: v, parent })
-      }${token(v.separator)}`;
-    }
-    function iterable_like(it, parent) {
-      return ts.definition(ts.wrap([
-        extended_attributes(it.extAttrs),
-        token(it.tokens.readonly),
-        token(it.tokens.base),
-        token(it.tokens.open),
-        ts.wrap(it.idlType.map(type)),
-        token(it.tokens.close),
-        token(it.tokens.termination)
-      ]), { data: it, parent });
-    }
-    function callbackInterface(it) {
-      return ts.definition(
-        wrap`${ts.trivia(it.trivia.callback)}callback${container(it)}`,
-        { data: it }
-      );
-    }
-    function eof(it) {
-      return ts.trivia(it.trivia);
-    }
-
-    const table = {
-      interface: container,
-      "interface mixin": container,
-      namespace: container,
-      operation,
-      attribute,
-      dictionary: container,
-      field,
-      const: const_,
-      typedef,
-      includes,
-      callback,
-      enum: enum_,
-      "enum-value": enum_value,
-      iterable: iterable_like,
-      legacyiterable: iterable_like,
-      maplike: iterable_like,
-      setlike: iterable_like,
-      "callback interface": callbackInterface,
-      eof
-    };
-    function dispatch(it, parent) {
-      const dispatcher = table[it.type];
-      if (!dispatcher) {
-        throw new Error(`Type "${it.type}" is unsupported`);
-      }
-      return table[it.type](it, parent);
-    }
-    function iterate(things, parent) {
-      if (!things) return;
-      const results = things.map(thing => dispatch(thing, parent));
-      return ts.wrap(results);
-    }
-    return iterate(ast);
+    return table[it.type](it, parent);
   }
-
-  const obj = {
-    write
-  };
-
-  if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
-    module.exports = obj;
-  } else if (typeof define === 'function' && define.amd) {
-    define([], () => obj);
-  } else {
-    (self || window).WebIDL2Writer = obj;
+  function iterate(things, parent) {
+    if (!things) return;
+    const results = things.map(thing => dispatch(thing, parent));
+    return ts.wrap(results);
   }
-})();
+  return iterate(ast);
+}
