+++ package.json
@@ -14,16 +14,18 @@
     "eslint": "^5.9.0",
     "expect": "23.6.0",
     "jsondiffpatch": "0.3.11",
-    "mocha": "5.2.0"
+    "mocha": "5.2.0",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0"
   },
   "scripts": {
     "lint": "eslint lib/*.js test/*.js test/util/*.js",
-    "test": "npm run lint && mocha",
+    "test": "npm run lint && webpack && mocha",
     "acquire": "node test/util/acquire.js"
   },
   "repository": "git://github.com/w3c/webidl2.js",
-  "main": "index.js",
+  "main": "dist/webidl2.js",
   "files": [
-    "lib/*"
+    "dist/*"
   ]
 }
