+++ webpack.config.js
@@ -1,0 +1,15 @@
+const path = require("path");
+
+module.exports = {
+  entry: "./index.js",
+  output: {
+    filename: "webidl2.js",
+    path: path.resolve(__dirname, "dist"),
+    library: "WebIDL2",
+    libraryExport: "default",
+    libraryTarget: "umd",
+    globalObject: "this"
+  },
+  mode: "production",
+  devtool: "source-map"
+};
